+++ packages/vue/components/row.js
@@ -10,20 +10,37 @@ export default {
     tag: {
       type: String,
       default: 'div'
+    },
+    resizable: Boolean,
+    resizableFixed: Boolean,
+    resizableAbsolute: Boolean,
+    resizableHandler: {
+      type: Boolean,
+      default: true
     }
   }, Mixins.colorProps),
-  render: function render() {
-    var _h = this.$createElement;
-    var self = this;
-    var props = self.props;
-    var className = props.className,
-        id = props.id,
-        style = props.style,
-        tag = props.tag,
-        noGap = props.noGap;
-    var RowTag = tag;
-    var classes = Utils.classNames(className, 'row', {
-      'no-gap': noGap
+
+  render() {
+    const _h = this.$createElement;
+    const self = this;
+    const props = self.props;
+    const {
+      className,
+      id,
+      style,
+      tag,
+      noGap,
+      resizable,
+      resizableFixed,
+      resizableAbsolute,
+      resizableHandler
+    } = props;
+    const RowTag = tag;
+    const classes = Utils.classNames(className, 'row', {
+      'no-gap': noGap,
+      resizable,
+      'resizable-fixed': resizableFixed,
+      'resizable-absolute': resizableAbsolute
     }, Mixins.colorClasses(props));
     return _h(RowTag, {
       style: style,
@@ -32,32 +49,53 @@ export default {
       attrs: {
         id: id
       }
-    }, [this.$slots['default']]);
+    }, [this.$slots['default'], resizable && resizableHandler && _h('span', {
+      class: 'resize-handler'
+    })]);
   },
-  created: function created() {
-    Utils.bindMethods(this, ['onClick']);
+
+  created() {
+    Utils.bindMethods(this, ['onClick', 'onResize']);
   },
-  mounted: function mounted() {
-    this.$refs.el.addEventListener('click', this.onClick);
+
+  mounted() {
+    const self = this;
+    self.eventTargetEl = self.$refs.el;
+    self.eventTargetEl.addEventListener('click', self.onClick);
+    self.$f7ready(f7 => {
+      f7.on('gridResize', self.onResize);
+    });
   },
-  beforeDestroy: function beforeDestroy() {
-    this.$refs.el.removeEventListener('click', this.onClick);
+
+  beforeDestroy() {
+    const self = this;
+    const el = self.$refs.el;
+    if (!el || !self.$f7) return;
+    el.removeEventListener('click', self.onClick);
+    self.$f7.off('gridResize', self.onResize);
+    delete self.eventTargetEl;
   },
+
   methods: {
-    onClick: function onClick(event) {
+    onClick(event) {
       this.dispatchEvent('click', event);
     },
-    dispatchEvent: function dispatchEvent(events) {
-      for (var _len = arguments.length, args = new Array(_len > 1 ? _len - 1 : 0), _key = 1; _key < _len; _key++) {
-        args[_key - 1] = arguments[_key];
-      }
 
-      __vueComponentDispatchEvent.apply(void 0, [this, events].concat(args));
+    onResize(el) {
+      if (el === this.eventTargetEl) {
+        this.dispatchEvent('grid:resize gridResize');
+      }
+    },
+
+    dispatchEvent(events, ...args) {
+      __vueComponentDispatchEvent(this, events, ...args);
     }
+
   },
   computed: {
-    props: function props() {
+    props() {
       return __vueComponentProps(this);
     }
+
   }
 };