+++ src/core/modules/router/back.js
@@ -115,6 +115,9 @@ function backward(el, backwardOptions) {
       .addClass(`navbar-previous${isMaster ? ' navbar-master' : ''}${isDetail ? ' navbar-master-detail' : ''}${isDetailRoot ? ' navbar-master-detail-root' : ''}`)
       .removeClass('stacked')
       .removeAttr('aria-hidden');
+    if (isMaster || isDetailRoot) {
+      router.emit('navbarRole', $newNavbarEl[0], { role: isMaster ? 'master' : 'detail', detailRoot: !!isDetailRoot });
+    }
   }
 
   // Remove previous page in case of "forced"
@@ -239,6 +242,7 @@ function backward(el, backwardOptions) {
       router.emit('pageMasterUnstack', $newPage[0]);
       if (dynamicNavbar) {
         $(app.navbar.getElByPage($newPage)).removeClass('navbar-master-stacked');
+        router.emi('navbarMasterUnstack', app.navbar.getElByPage($newPage));
       }
     }
     // Page init and before init events
@@ -361,15 +365,11 @@ function backward(el, backwardOptions) {
   // Animation
   function afterAnimation() {
     // Set classes
-    const pageClasses = 'page-previous page-current page-next';
-    const navbarClasses = 'navbar-previous navbar-current navbar-next';
-    $newPage.removeClass(pageClasses).addClass('page-current').removeAttr('aria-hidden').trigger('page:position', { position: 'current' });
-    router.emit('pagePosition', $newPage[0], 'current');
-    $oldPage.removeClass(pageClasses).addClass('page-next').attr('aria-hidden', 'true').trigger('page:position', { position: 'next' });
-    router.emit('pagePosition', $oldPage[0], 'next');
+    router.setPagePosition($newPage, 'current', false);
+    router.setPagePosition($oldPage, 'next', true);
     if (dynamicNavbar) {
-      $newNavbarEl.removeClass(navbarClasses).addClass('navbar-current').removeAttr('aria-hidden');
-      $oldNavbarEl.removeClass(navbarClasses).addClass('navbar-next').attr('aria-hidden', 'true');
+      router.setNavbarPosition($newNavbarEl, 'current', false);
+      router.setNavbarPosition($oldNavbarEl, 'next', true);
     }
 
     // After animation event
@@ -406,15 +406,11 @@ function backward(el, backwardOptions) {
   }
 
   function setPositionClasses() {
-    const pageClasses = 'page-previous page-current page-next';
-    const navbarClasses = 'navbar-previous navbar-current navbar-next';
-    $oldPage.removeClass(pageClasses).addClass('page-current').trigger('page:position', { position: 'current' });
-    router.emit('pagePosition', $oldPage[0], 'current');
-    $newPage.removeClass(pageClasses).addClass('page-previous').removeAttr('aria-hidden').trigger('page:position', { position: 'previous' });
-    router.emit('pagePosition', $newPage[0], 'previous');
+    router.setPagePosition($oldPage, 'current');
+    router.setPagePosition($newPage, 'previous', false);
     if (dynamicNavbar) {
-      $oldNavbarEl.removeClass(navbarClasses).addClass('navbar-current');
-      $newNavbarEl.removeClass(navbarClasses).addClass('navbar-previous').removeAttr('aria-hidden');
+      router.setNavbarPosition($oldNavbarEl, 'current');
+      router.setNavbarPosition($newNavbarEl, 'previous', false);
     }
   }