+++ src/core/modules/router/navigate.js
@@ -187,6 +187,9 @@ function forward(el, forwardOptions = {}) {
       .removeClass('navbar-previous navbar-current navbar-next')
       .addClass(`navbar-${newPagePosition}${isMaster ? ' navbar-master' : ''}${isDetail ? ' navbar-master-detail' : ''}${isDetailRoot ? ' navbar-master-detail-root' : ''}`)
       .removeClass('stacked');
+    if (isMaster || isDetail) {
+      router.emit('navbarRole', $newNavbarEl[0], { role: isMaster ? 'master' : 'detail', detailRoot: !!isDetailRoot });
+    }
   }
 
   // Find Old Page
@@ -208,6 +211,8 @@ function forward(el, forwardOptions = {}) {
       $oldNavbarEl = $navbarsInView.filter((index, navbarEl) => navbarEl !== $newNavbarEl[0]);
     }
   } else {
+    let removedPageEls = [];
+    let removedNavbarEls = [];
     if ($pagesInView.length > 1) {
       let i = 0;
       for (i = 0; i < $pagesInView.length - 1; i += 1) {
@@ -219,6 +224,7 @@ function forward(el, forwardOptions = {}) {
           router.emit('pageMasterStack', $pagesInView[i]);
           if (dynamicNavbar) {
             $(app.navbar.getElByPage(masterPageEl)).addClass('navbar-master-stacked');
+            router.emit('navbarMasterStack', app.navbar.getElByPage(masterPageEl));
           }
           continue; // eslint-disable-line
         }
@@ -232,9 +238,11 @@ function forward(el, forwardOptions = {}) {
           }
         } else {
           // Page remove event
+          removedPageEls.push($pagesInView[i]);
           router.pageCallback('beforeRemove', $pagesInView[i], $navbarsInView && $navbarsInView[i], 'previous', undefined, options);
           router.removePage($pagesInView[i]);
           if (dynamicNavbar && oldNavbarEl) {
+            removedNavbarEls.push(oldNavbarEl);
             router.removeNavbar(oldNavbarEl);
           }
         }
@@ -242,12 +250,14 @@ function forward(el, forwardOptions = {}) {
     }
     $oldPage = $viewEl
       .children('.page:not(.stacked)')
-      .filter((index, page) => page !== $newPage[0]);
+      .filter((index, pageEl) => pageEl !== $newPage[0] && removedPageEls.indexOf(pageEl) < 0);
     if (dynamicNavbar) {
       $oldNavbarEl = $navbarsEl
         .children('.navbar:not(.stacked)')
-        .filter((index, navbarEl) => navbarEl !== $newNavbarEl[0]);
+        .filter((index, navbarEl) => navbarEl !== $newNavbarEl[0] && removedNavbarEls.indexOf(removedNavbarEls) < 0);
     }
+    removedPageEls = [];
+    removedNavbarEls = [];
   }
 
   if (isDetail && !options.reloadAll) {
@@ -438,14 +448,9 @@ function forward(el, forwardOptions = {}) {
     router.pageCallback('afterIn', $newPage, $newNavbarEl, newPagePosition, 'current', options);
     if (options.reloadCurrent && options.clearPreviousHistory) router.clearPreviousHistory();
     if (reloadDetail) {
-      masterPageEl.classList.add('page-previous');
-      masterPageEl.classList.remove('page-current');
-      $(masterPageEl).trigger('page:position', { position: 'previous' });
-      router.emit('pagePosition', masterPageEl, 'previous');
-
+      router.setPagePosition($(masterPageEl), 'previous');
       if (masterPageEl.f7Page && masterPageEl.f7Page.navbarEl) {
-        masterPageEl.f7Page.navbarEl.classList.add('navbar-previous');
-        masterPageEl.f7Page.navbarEl.classList.remove('navbar-current');
+        router.setNavbarPosition($(masterPageEl.f7Page.navbarEl), 'previous');
       }
     }
     return router;
@@ -461,22 +466,11 @@ function forward(el, forwardOptions = {}) {
 
   // Animation
   function afterAnimation() {
-    const pageClasses = 'page-previous page-current page-next';
-    const navbarClasses = 'navbar-previous navbar-current navbar-next';
-    $newPage.removeClass(pageClasses).addClass('page-current').removeAttr('aria-hidden').trigger('page:position', { position: 'current' });
-    router.emit('pagePosition', $newPage[0], 'current');
-    $oldPage.removeClass(pageClasses).addClass('page-previous').trigger('page:position', { position: 'previous' });
-    router.emit('pagePosition', $oldPage[0], 'previous');
-
-    if (!$oldPage.hasClass('page-master')) {
-      $oldPage.attr('aria-hidden', 'true');
-    }
+    router.setPagePosition($newPage, 'current', false);
+    router.setPagePosition($oldPage, 'previous', !$oldPage.hasClass('page-master'));
     if (dynamicNavbar) {
-      $newNavbarEl.removeClass(navbarClasses).addClass('navbar-current').removeAttr('aria-hidden');
-      $oldNavbarEl.removeClass(navbarClasses).addClass('navbar-previous');
-      if (!$oldNavbarEl.hasClass('navbar-master')) {
-        $oldNavbarEl.attr('aria-hidden', 'true');
-      }
+      router.setNavbarPosition($newNavbarEl, 'current', false);
+      router.setNavbarPosition($oldNavbarEl, 'previous', !$oldNavbarEl.hasClass('navbar-master'));
     }
     // After animation event
     router.allowPageChange = true;
@@ -514,15 +508,11 @@ function forward(el, forwardOptions = {}) {
     }
   }
   function setPositionClasses() {
-    const pageClasses = 'page-previous page-current page-next';
-    const navbarClasses = 'navbar-previous navbar-current navbar-next';
-    $oldPage.removeClass(pageClasses).addClass('page-current').removeAttr('aria-hidden').trigger('page:position', { position: 'current' });
-    router.emit('pagePosition', $oldPage[0], 'current');
-    $newPage.removeClass(pageClasses).addClass('page-next').removeAttr('aria-hidden').trigger('page:position', { position: 'next' });
-    router.emit('pagePosition', $newPage[0], 'next');
+    router.setPagePosition($oldPage, 'current', false);
+    router.setPagePosition($newPage, 'next', false);
     if (dynamicNavbar) {
-      $oldNavbarEl.removeClass(navbarClasses).addClass('navbar-current').removeAttr('aria-hidden');
-      $newNavbarEl.removeClass(navbarClasses).addClass('navbar-next').removeAttr('aria-hidden');
+      router.setNavbarPosition($oldNavbarEl, 'current', false);
+      router.setNavbarPosition($newNavbarEl, 'next', false);
     }
   }
   if (options.animate && !(isMaster && app.width >= router.params.masterDetailBreakpoint)) {