+++ src/phenome/components/navbar.jsx
@@ -44,6 +44,11 @@ export default {
     }
     return {
       _theme: $f7 ? self.$theme : null,
+      routerPositionClass: '',
+      largeCollapsed: false,
+      routerNavbarRole: null,
+      routerNavbarRoleDetailRoot: false,
+      routerNavbarMasterStack: false,
     };
   },
   render() {
@@ -70,7 +75,11 @@ export default {
       titleLarge,
     } = props;
 
-    const theme = self.state.theme;
+    const {
+      _theme: theme,
+      routerPositionClass,
+      largeCollapsed,
+    } = self.state;
 
     let leftEl;
     let titleEl;
@@ -86,10 +95,18 @@ export default {
     const classes = Utils.classNames(
       className,
       'navbar',
+      routerPositionClass,
       {
         'navbar-hidden': hidden,
         'navbar-large': large,
         'navbar-large-transparent': largeTransparent,
+        'navbar-large-collapsed': large && largeCollapsed,
+        'navbar-master': this.state.routerNavbarRole === 'master',
+        'navbar-master-detail': this.state.routerNavbarRole === 'detail',
+        'navbar-master-detail-root': this.state.routerNavbarRoleDetailRoot === true,
+        'navbar-master-stacked': this.state.routerNavbarMasterStack === true,
+        'no-shadow': noShadow,
+        'no-hairline': noHairline,
       },
       Mixins.colorClasses(props),
     );
@@ -138,8 +155,6 @@ export default {
           innerClassName,
           {
             sliding,
-            'no-shadow': noShadow,
-            'no-hairline': noHairline,
             'navbar-inner-left-title': addLeftTitleClass,
             'navbar-inner-centered-title': addCenterTitleClass,
           }
@@ -163,7 +178,7 @@ export default {
     );
   },
   componentDidCreate() {
-    Utils.bindMethods(this, ['onBackClick', 'onHide', 'onShow', 'onExpand', 'onCollapse']);
+    Utils.bindMethods(this, ['onBackClick', 'onHide', 'onShow', 'onExpand', 'onCollapse', 'onNavbarPosition']);
   },
   componentDidMount() {
     const self = this;
@@ -175,6 +190,10 @@ export default {
       f7.on('navbarHide', self.onHide);
       f7.on('navbarCollapse', self.onCollapse);
       f7.on('navbarExpand', self.onExpand);
+      f7.on('navbarPosition', self.onNavbarPosition);
+      f7.on('navbarRole', self.onNavbarRole);
+      f7.on('navbarMasterStack', self.onNavbarMasterStack);
+      f7.on('navbarMasterUnstack', self.onNavbarMasterUnstack);
     });
   },
   componentDidUpdate() {
@@ -192,6 +211,10 @@ export default {
     f7.off('navbarHide', self.onHide);
     f7.off('navbarCollapse', self.onCollapse);
     f7.off('navbarExpand', self.onExpand);
+    f7.off('navbarPosition', self.onNavbarPosition);
+    f7.off('navbarRole', self.onNavbarRole);
+    f7.off('navbarMasterStack', self.onNavbarMasterStack);
+    f7.off('navbarMasterUnstack', self.onNavbarMasterUnstack);
     self.eventTargetEl = null;
     delete self.eventTargetEl;
   },
@@ -206,11 +229,42 @@ export default {
     },
     onExpand(navbarEl) {
       if (this.eventTargetEl !== navbarEl) return;
+      this.setState({
+        largeCollapsed: false,
+      });
       this.dispatchEvent('navbar:expand navbarExpand');
     },
     onCollapse(navbarEl) {
       if (this.eventTargetEl !== navbarEl) return;
+      this.setState({
+        largeCollapsed: true,
+      });
       this.dispatchEvent('navbar:collapse navbarCollapse');
+    },
+    onNavbarPosition(navbarEl, position) {
+      if (this.eventTargetEl !== navbarEl) return;
+      this.setState({
+        routerPositionClass: `navbar-${position}`,
+      });
+    },
+    onNavbarRole(navbarEl, rolesData) {
+      if (this.eventTargetEl !== navbarEl) return;
+      this.setState({
+        routerNavbarRole: rolesData.role,
+        routerNavbarRoleDetailRoot: rolesData.detailRoot,
+      });
+    },
+    onNavbarMasterStack(navbarEl) {
+      if (this.eventTargetEl !== navbarEl) return;
+      this.setState({
+        routerNavbarMasterStack: true,
+      });
+    },
+    onNavbarMasterUnstack(navbarEl) {
+      if (this.eventTargetEl !== navbarEl) return;
+      this.setState({
+        routerNavbarMasterStack: false,
+      });
     },
     hide(animate) {
       const self = this;